+++ packages/create-sitecore-jss/src/templates/nextjs-sxa/src/components/Navigation.tsx
@@ -2,6 +2,7 @@ import { useState, JSX } from 'react';
 import {
   Link,
   LinkField,
+  FieldMetadata,
   Text,
   TextField,
   useSitecoreContext,
@@ -39,12 +40,13 @@ const getNavigationText = function(props: NavigationProps): JSX.Element | string
   return text;
 };
 
-const getLinkField = (props: NavigationProps): LinkField => ({
+const getLinkField = (props: NavigationProps): LinkField & FieldMetadata => ({
   value: {
     href: props.fields.Href,
     title: getLinkTitle(props),
     querystring: props.fields.Querystring,
   },
+  metadata: props.fields.NavigationTitle?.metadata,
 });
 
 export const Default = (props: NavigationProps): JSX.Element => {