+++ packages/sitecore-jss-nextjs/src/index.ts
@@ -28,6 +28,7 @@ export {
   PlaceholdersData,
   RouteData,
   Field,
+  FieldMetadata,
   Item,
   HtmlElementRendering,
   getChildPlaceholder,